+++ frame/lending/src/mocks_offchain/mod.rs
@@ -259,7 +259,7 @@ parameter_types! {
 	pub const MinU32: u32 = 0;
 	pub const MinU64: u64 = 0;
 
-	pub const TwapWindow: u32 = 3;
+	pub const TwapWindow: u16 = 3;
 }
 
 pub struct Decimals;
